Fix destruction of exercise not looking for states in server side

+++ src/app/Controllers/ExerciseManagementController.php
@@ -13,20 +13,26 @@ class ExerciseManagementController extends Controller
         $exercisesAnswer = Exercises::where('state_id', States::slug('ANSWER'))->get();
         $exercisesClose  = Exercises::where('state_id', States::slug('CLOSE'))->get();
 
-        return $this->render('exercise-management', [
-            'exercisesBuild'  => $exercisesBuild,
-            'exercisesAnswer' => $exercisesAnswer, 'exercisesClose' => $exercisesClose
-        ], 
-        [
-            'description' => 'List of all exercises and their states',
-            'keywords'    => 'Exercise, Exercise Management, Exercise List'
-        ]);
+        return $this->render(
+            'exercise-management',
+            [
+                'exercisesBuild'  => $exercisesBuild,
+                'exercisesAnswer' => $exercisesAnswer, 'exercisesClose' => $exercisesClose
+            ],
+            [
+                'description' => 'List of all exercises and their states',
+                'keywords'    => 'Exercise, Exercise Management, Exercise List'
+            ]
+        );
     }
 
     public function removeExercise($params)
     {
         $exercise = Exercises::find($params['id']);
-        $exercise->delete();
+
+        if ($exercise->isBuild() || $exercise->isClosed()) {
+            $exercise->delete();
+        }
 
         header("Location: /exercises");
         exit();
